+++ air/tests/test_module/negative_tests/preparation_step.rs
@@ -40,13 +40,13 @@ fn invalid_data_without_versions() {
     let script = r#"(null)"#;
     let invalid_data = InvalidDataStruct { trace: vec![1, 2, 3] };
 
-    let invalid_data = InterpreterDataRepr.get_format().to_vec(&invalid_data).unwrap();
+    let invalid_data = InterpreterDataEnvelopeRepr.get_format().to_vec(&invalid_data).unwrap();
 
     let result = call_vm!(vm, <_>::default(), script, "", invalid_data.clone());
 
-    let expected_serde_error = InterpreterData::try_from_slice(&invalid_data).unwrap_err();
-    let expected_error = PreparationError::DataDeFailed {
-        data: invalid_data,
+    let expected_serde_error = InterpreterDataEnvelope::try_from_slice(&invalid_data).unwrap_err();
+    let expected_error = PreparationError::EnvelopeDeFailed {
+        env_raw_data: invalid_data,
         error: expected_serde_error,
     };
     assert!(check_error(&result, expected_error));
@@ -73,13 +73,13 @@ fn invalid_data_with_versions() {
         trace: vec![1, 2, 3],
         versions: versions.clone(),
     };
-    let invalid_data = InterpreterDataRepr.get_format().to_vec(&invalid_data).unwrap();
+    let invalid_data = InterpreterDataEnvelopeRepr.get_format().to_vec(&invalid_data).unwrap();
 
     let result = call_vm!(vm, <_>::default(), script, "", invalid_data.clone());
 
-    let expected_serde_error = InterpreterData::try_from_slice(&invalid_data).unwrap_err();
-    let expected_error = PreparationError::DataDeFailedWithVersions {
-        data: invalid_data,
+    let expected_serde_error = InterpreterDataEnvelope::try_from_slice(&invalid_data).unwrap_err();
+    let expected_error = PreparationError::EnvelopeDeFailedWithVersions {
+        env_raw_data: invalid_data,
         error: expected_serde_error,
         versions,
     };
@@ -92,7 +92,7 @@ fn invalid_callresults() {
 
     let air = r#"(null)"#.to_string();
     let client_peer_id = "some_peer_id".to_string();
-    let prev_data = InterpreterData::new(semver::Version::new(1, 1, 1));
+    let prev_data = InterpreterDataEnvelope::new(semver::Version::new(1, 1, 1));
     let prev_data: Vec<u8> = prev_data.serialize().unwrap();
     let data = Vec::<u8>::new();
     let vec = Vec::<u8>::new();