+++ junk/gen-bench-data/src/cid_benchmarking/mod.rs
@@ -110,8 +110,6 @@ pub fn cid_benchmarking_data(
     peer_id: String,
     particle_id: &str,
 ) -> Vec<u8> {
-    use air_interpreter_sede::Format;
-
     let mut curr_data: PreCidInterpeterData = read_data(curr_data_filename);
     let calls: TraceCalls = read_data("src/cid_benchmarking/simple-calls-info.json");
     let mut calls = calls.into_iter();
@@ -161,7 +159,19 @@ pub fn cid_benchmarking_data(
         .unwrap()
         .insert("signatures".to_owned(), json!(ss));
 
-    InterpreterDataFormat::default().to_vec(&curr_data).unwrap().into()
+    let to_value = serde_json::to_value(curr_data).unwrap();
+    let inner_data = serde_json::from_value::<InterpreterData>(to_value)
+        .unwrap()
+        .serialize()
+        .unwrap()
+        .into();
+
+    let data_env = InterpreterDataEnvelope {
+        versions: Versions::new(interpreter_version().clone()),
+        inner_data,
+    };
+
+    data_env.serialize().unwrap()
 }
 
 pub fn cid_benchmarking_long_data(