+++ air-interpreter/CHANGELOG.md
@@ -1,5 +1,54 @@
 # Changelog
 
+## [0.58.0](https://github.com/fluencelabs/aquavm/compare/air-interpreter-v0.57.0...air-interpreter-v0.58.0) (2024-01-11)
+
+
+### Features
+
+* **air,air-cli:** pretty-printing binary interpreter data ([#794](https://github.com/fluencelabs/aquavm/issues/794)) ([d6b1da9](https://github.com/fluencelabs/aquavm/commit/d6b1da9bdc1197e72ef24051293fd06d3842f318))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * aquavm-air bumped from 0.57.0 to 0.58.0
+    * air-interpreter-interface bumped from 0.17.0 to 0.17.1
+
+## [0.57.0](https://github.com/fluencelabs/aquavm/compare/air-interpreter-v0.56.0...air-interpreter-v0.57.0) (2024-01-09)
+
+
+### Miscellaneous Chores
+
+* **air-interpreter:** Synchronize air-interpreter versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * aquavm-air bumped from 0.56.0 to 0.57.0
+    * air-interpreter-interface bumped from 0.16.0 to 0.17.0
+
+## [0.56.0](https://github.com/fluencelabs/aquavm/compare/air-interpreter-v0.55.0...air-interpreter-v0.56.0) (2024-01-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **data:** flexible serialization formats ([#757](https://github.com/fluencelabs/aquavm/issues/757))
+
+### Features
+
+* **data:** flexible serialization formats ([#757](https://github.com/fluencelabs/aquavm/issues/757)) ([771d42d](https://github.com/fluencelabs/aquavm/commit/771d42dec43d3081621897edda3735768fd9ff71))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * aquavm-air bumped from 0.55.0 to 0.56.0
+    * air-interpreter-interface bumped from 0.15.2 to 0.16.0
+
 ## [0.55.0](https://github.com/fluencelabs/aquavm/compare/air-interpreter-v0.54.0...air-interpreter-v0.55.0) (2023-12-12)